move transcription example to dedicated file
